Restructure backend directory and migrate database configuration
- Added DB_PORT to .env for SQL Server connection.
- Moved BED_CareConnectDB.sql to  the backend directory.
- Moved dbConfig.js to backend/db and updated references in app.js.
- Reorganized controllers, models, middlewares, and routes for medication to backend/functions folder directory.
- Updated medication and reminder functionalities to align with the new structure.
- Updated dotenv package to version 17.1.0 for enhanced security features.
