+++ elderly-care-app/.env
@@ -3,3 +3,4 @@ DB_USER=careconnect_user
 DB_PASSWORD=careconnect
 DB_SERVER=localhost
 DB_NAME=CareConnectDB
+DB_PORT=1433
