+++ elderly-care-app/backend/db/BED_CareConnectDB.sql
