+++ elderly-care-app/backend/db/dbConfig.js
