+++ elderly-care-app/backend/functions/medication/controllers/medicationController.js
