+++ elderly-care-app/backend/functions/medication/middlewares/validateMedication.js
