+++ elderly-care-app/package.json
@@ -11,7 +11,7 @@
   "description": "",
   "dependencies": {
     "bcrypt": "^6.0.0",
-    "dotenv": "^17.0.1",
+    "dotenv": "^17.1.0",
     "express": "^5.1.0",
     "joi": "^17.13.3",
     "jsonwebtoken": "^9.0.2",
